Fix the Java Integer types on Windows

+++ src/java.base/windows/native/include/jni_md.h
@@ -32,9 +32,8 @@
 #define JNIIMPORT __declspec(dllimport)
 #define JNICALL __stdcall
 
-// 'long' is always 32 bit on windows so this matches what jdk expects
-typedef long jint;
-typedef __int64 jlong;
+typedef int jint;
+typedef long long jlong;
 typedef signed char jbyte;
 
 #endif /* !_JAVASOFT_JNI_MD_H_ */
